Rename ReactTypeOfSideEffect too
