+++ packages/react-reconciler/src/ReactUpdateQueue.js
@@ -88,11 +88,7 @@ import type {Fiber} from './ReactFiber';
 import type {ExpirationTime} from './ReactFiberExpirationTime';
 
 import {NoWork} from './ReactFiberExpirationTime';
-import {
-  Callback,
-  ShouldCapture,
-  DidCapture,
-} from 'shared/ReactTypeOfSideEffect';
+import {Callback, ShouldCapture, DidCapture} from 'shared/ReactSideEffectTags';
 import {ClassComponent, ClassComponentLazy} from 'shared/ReactWorkTags';
 
 import {
